@nsajko 2024-05-08
The `StaticView` type is a simplified version of my unregistered package StaticViews.jl. This PR was prompted by @jishnub recent message on Slack about `in(::Any, ::Tuple)`.