@nsajko 2024-05-08
The benefit of this approach, with `StaticView`, is that it avoids calling `tail` on a `Tuple`.