@jishnub 2024-05-08
We may want a cutoff length for this, as the present loop-based implementation is faster for large tuples.
On nightly
```julia
julia> t = ntuple(x->'A', 1000);

julia> @btime 'C' in $t;
  354.118 ns (0 allocations: 0 bytes)
```
This PR
```julia
julia> @btime 'C' in $t;
  2.889 μs (0 allocations: 0 bytes)
```