@nsajko 2024-05-08
Taking the example from Slack with constant folding, this is with this PR applied:

```julia-repl
julia> struct WrapperChar <: AbstractChar
           wrapperchar :: Char
           isuppertri :: Bool
       end

julia> function Base.Char(w::WrapperChar)
           T = w.wrapperchar
           if T ∉ ('S', 'H')
               T
           else
               _isuppertri(w) ? uppercase(T) : lowercase(T)
           end
       end

julia> Base.codepoint(w::WrapperChar) = codepoint(Char(w))

julia> WrapperChar(n::UInt32) = WrapperChar(Char(n), true)
WrapperChar

julia> _isuppertri(w::WrapperChar) = w.isuppertri
_isuppertri (generic function with 1 method)

julia> @code_typed (() -> in(WrapperChar('N'), ('A', 'B', 'C')))()
CodeInfo(
1 ─     return false
) => Bool
```

With nightly:

```julia-repl
julia> @code_typed (() -> in(WrapperChar('N'), ('A', 'B', 'C')))()
CodeInfo(
1 ──       nothing::Nothing
2 ┄─ %2  = φ (#1 => 'A', #11 => %21)::Char
│    %3  = φ (#1 => 2, #11 => %22)::Int64
│    %4  = Base.bitcast(Base.UInt32, %2)::UInt32
│    %5  = Base.bitcast(Base.UInt32, 'N')::UInt32
│    %6  = (%4 === %5)::Bool
└───       goto #4 if not %6
3 ──       goto #13
4 ── %9  = Base.sle_int(1, %3)::Bool
└───       goto #6 if not %9
5 ── %11 = Base.sle_int(%3, 3)::Bool
└───       goto #7
6 ──       nothing::Nothing
7 ┄─ %14 = φ (#5 => %11, #6 => false)::Bool
└───       goto #9 if not %14
8 ── %16 = $(Expr(:boundscheck, true))::Bool
│    %17 = Base.getfield(('A', 'B', 'C'), %3, %16)::Char
│    %18 = Base.add_int(%3, 1)::Int64
└───       goto #10
9 ──       goto #10
10 ┄ %21 = φ (#8 => %17)::Char
│    %22 = φ (#8 => %18)::Int64
│    %23 = φ (#8 => false, #9 => true)::Bool
│    %24 = Base.not_int(%23)::Bool
└───       goto #12 if not %24
11 ─       goto #2
12 ─       goto #13
13 ┄ %28 = φ (#3 => true, #12 => false)::Bool
└───       return %28
) => Bool
```