@aviatesk 2024-05-09
Can you explain the advantages of using `StaticView`, especially how it makes it unnecessary to use `tail`? I'm particularly interested in its advantages compared to the technique that uses the length of argument tuple as a threshold to split dispatches between recursive and loop implementations, like in the implementation of `map`.