@nsajko 2024-05-09
This is not satisfactory:

```julia-repl
julia> Core.Compiler.return_type(∈, Tuple{Int,Tuple{Int,Int}})
Union{Missing, Bool}
```