@nsajko 2024-05-10
Performance comparison show 3x -4x improvement:

Before:

```julia-repl
julia> using BenchmarkTools

julia> t = ntuple(Returns(7), 100);

julia> minimum(@benchmark 3 ∈ $t)
BenchmarkTools.TrialEstimate: 
  time:             34.215 ns
  gctime:           0.000 ns (0.00%)
  memory:           0 bytes
  allocs:           0

julia> t = ntuple(Returns(7), 500);

julia> minimum(@benchmark 3 ∈ $t)
BenchmarkTools.TrialEstimate: 
  time:             203.996 ns
  gctime:           0.000 ns (0.00%)
  memory:           0 bytes
  allocs:           0

julia> versioninfo()
Julia Version 1.12.0-DEV.495
Commit 3c966a5107a (2024-05-09 11:53 UTC)
Build Info:
  Official https://julialang.org/ release
Platform Info:
  OS: Linux (x86_64-linux-gnu)
  CPU: 8 × AMD Ryzen 3 5300U with Radeon Graphics
  WORD_SIZE: 64
  LLVM: libLLVM-17.0.6 (ORCJIT, znver2)
Threads: 1 default, 0 interactive, 1 GC (on 8 virtual cores)
```

After:

```julia-repl
julia> using BenchmarkTools

julia> t = ntuple(Returns(7), 100);

julia> minimum(@benchmark 3 ∈ $t)
BenchmarkTools.TrialEstimate:
  time:             9.829 ns
  gctime:           0.000 ns (0.00%)
  memory:           0 bytes
  allocs:           0

julia> t = ntuple(Returns(7), 500);

julia> minimum(@benchmark 3 ∈ $t)
BenchmarkTools.TrialEstimate:
  time:             53.974 ns
  gctime:           0.000 ns (0.00%)
  memory:           0 bytes
  allocs:           0

julia> versioninfo()
Julia Version 1.12.0-DEV.502
Commit 9598e8dce1 (2024-05-10 07:13 UTC)
Platform Info:
  OS: Linux (x86_64-pc-linux-gnu)
  CPU: 8 × AMD Ryzen 3 5300U with Radeon Graphics
  WORD_SIZE: 64
  LLVM: libLLVM-17.0.6 (ORCJIT, znver2)
Threads: 1 default, 0 interactive, 1 GC (on 8 virtual cores)
```