@aviatesk 2024-05-10
The benchmark isn't fair because this PR increases the threshold for tuple lengths that permit unrolling. Essentially, the performance difference we're seeing is between the loop implementation (from the master) and the recursive implementation (in this PR) for the very large (100-length) tuples. If comparisons are needed, they should be made between recursive implementations. Something like this:
```julia
julia> @benchmark i ∈ t setup=(n = rand((2:10...,25,100)); t = ntuple(n) do _; rand(1:10); end; i = rand(1:10))
BenchmarkTools.Trial: 10000 samples with 998 evaluations.
 Range (min … max):  17.117 ns … 83.709 ns  ┊ GC (min … max): 0.00% … 0.00%
 Time  (median):     22.253 ns              ┊ GC (median):    0.00%
 Time  (mean ± σ):   22.440 ns ±  2.913 ns  ┊ GC (mean ± σ):  0.00% ± 0.00%

            ▁▁   ▃▂▄▁▁▂█▇▄                                     
  ▁▁▂▂▂▄▄▄▃▅██▅▅▇█████████▇▇▇▇▄▃▃▅▅▃▃▃▄▃▂▂▂▂▂▂▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁ ▃
  17.1 ns         Histogram: frequency by time        31.6 ns <

 Memory estimate: 0 bytes, allocs estimate: 0.
```