@nsajko 2024-05-10
> The benchmark isn't fair because this PR increases the threshold for tuple lengths that permit unrolling.

Increasing the threshold is the point of this PR (now that the competing PR is already merged).

> If comparisons are needed, they should be made between recursive implementations.

That obviously makes this PR's perf wins even greater, by a lot. The recursive implementation on master allocates when given tuples larger than intended (because of relying on `tail(::Tuple)`):

```julia-repl
julia> f(x, t) = Base._in_tuple(x, t, false)
f (generic function with 1 method)

julia> using BenchmarkTools

julia> t = ntuple(Returns(7), 100);

julia> minimum(@benchmark f(3,$t))
BenchmarkTools.TrialEstimate: 
  time:             175.159 μs
  gctime:           0.000 ns (0.00%)
  memory:           37.48 KiB
  allocs:           69
```